// src/chat/gateways/path-types.ts
export enum ChatPathTypes {
  JOIN_ROOM = 'join-room',
  LEAVE_ROOM = 'leave-room',
  GET_CHAT = 'get-chat',
  SEND_MESSAGE = 'send-message',
  REMOVE_MESSAGE = 'remove-message',
  JOIN_CHAT = 'join-group',
  UPDATE_CHAT = 'update-group',
  ADD_TYPING = 'add-typing',
  REMOVE_TYPING = 'remove-typing',
  GET_TYPING = 'get-typing',
  UPDATE_CHATS = 'update-chats',
}